Create an expander utility for use with custom CSS.
_But do not_ allow o-expander css class names to be modified.
This way expander features can be used to create novel ui,
but things don't get messy when overriding o-expander css.

There is a clear distinction between the component o-expander
and the utility which can be used to make custom expanding UIs.
