+++ src/js/expander.js
@@ -1,274 +1,58 @@
-import viewport from 'o-viewport';
+import ExpanderUtility from './expander-utility';
 
-// Used to find o-expander ids.
-let count = 0;
+class Expander extends ExpanderUtility {
 
-class Expander {
 	/**
-	 * Class constructor.
+	 * o-expander constructor.
 	 * @param {HTMLElement} [oExpanderElement] - The component element in the DOM
 	 * @param {Object} [opts={}] - An options object for configuring the component
 	 */
 	constructor (oExpanderElement, opts) {
-		// Set expander state.
-		// 'expanded', 'collapsed', or 'null';
-		this._currentState = null;
-
-		// Get configurable options.
-		this.options = Object.assign({}, {
-			shrinkTo: 'height',
-			toggleState: 'all',
-			contentSelector: '.o-expander__content',
-			toggleSelector: '.o-expander__toggle',
-		}, opts || Expander._getDataAttributes(oExpanderElement));
-
-		// If the user has not configured toggle text for the expanded state,
-		// set it based on the "shrinkTo" option: "hide" when hiding collapsed
-		// items; "less" when obscuring by reducing the container height by a
-		// given value; "fewer" otherwise.
-		if (!this.options.expandedToggleText) {
-			switch (this.options.shrinkTo) {
-				case 'hidden':
-					this.options.expandedToggleText = 'hide';
-					break;
-				case 'height':
-					this.options.expandedToggleText = 'less';
-					break;
-				default:
-					this.options.expandedToggleText = 'fewer';
-					break;
+		super(oExpanderElement, Object.assign({
+			selectors: {
+				toggle: '.o-expander__toggle',
+				content: '.o-expander__content',
+				contentItem: '.o-expander__content > li',
+			},
+			classnames: {
+				initialised: 'o-expander__content--initialised',
+				inactive: 'o-expander__content--inactive',
+				expanded: 'o-expander__content--expanded',
+				collapsed: 'o-expander__content--collapsed',
+				collapsibleItem: 'o-expander__collapsible-item'
 			}
-		}
-
-		// If the user has not configured toggle text for the collapsed state,
-		// set it based on the "shrinkTo" option: "show" hiding collapsed items;
-		// or "more" when collapsing to a height.
-		if (!this.options.collapsedToggleText) {
-			this.options.collapsedToggleText = this.options.shrinkTo === 'hidden' ? 'show' : 'more';
-		}
-
-		// Selectors.
-		this.className = 'o-expander';
-		this.countSelector = 'o-expander > li';
-
-		// Elements.
-		this.oExpanderElement = oExpanderElement;
-		this.contentElement = this.oExpanderElement.querySelector(this.options.contentSelector);
-		this.toggles = [].slice.apply(this.oExpanderElement.querySelectorAll(this.options.toggleSelector));
-		if (!this.toggles.length) {
-			throw new Error(
-				'o-expander needs a toggle link or button.' +
-				`None were found for toggle selector "${this.options.toggleSelector}".`
-			);
-		}
-
-		// If `shrinkTo` is a number, cast to an actual number using the
-		// unary operator `+`.
-		if (!isNaN(this.options.shrinkTo)) {
-			this.options.shrinkTo = +this.options.shrinkTo;
-		}
-
-		// Set `aria-controls` on each toggle using expander ids.
-		this.id = this.contentElement.id;
-		if (!this.id) {
-			while (document.querySelector('#o-expander__toggle--' + count)) {
-				count++;
-			}
-			this.id = this.contentElement.id = 'o-expander__toggle--' + count;
-		}
-		this.toggles.forEach(toggle => toggle.setAttribute('aria-controls', this.id));
-
-		// Add a click event to each toggle.
-		this.toggles.forEach(toggle => {
-			toggle.addEventListener('click', () => this.toggle());
-		});
-
-		// If shrinking based on a height set in css, reapply the expander on
-		// orientation and resize events.
-		if (this.options.shrinkTo === 'height') {
-			viewport.listenTo('resize');
-			viewport.listenTo('orientation');
-			document.body.addEventListener('oViewport.orientation', () => this.apply());
-			document.body.addEventListener('oViewport.resize', () => this.apply());
-		}
-
-		// Add a data attribute to indicate the expander is initialised, which
-		// may be styled against for progressive enhancement (we shouldn't hide
-		// content when the expander fails to load).
-		this.oExpanderElement.setAttribute('data-o-expander-js', '');
-
-		// Apply the configured expander.
-		this.apply(true);
-
-		// Setup. Fire the `oExpander.init` event.
-		this._dispatchEvent('init');
+		}, opts || Expander._getDataAttributes(oExpanderElement)));
 	}
 
 	/**
-	 * Recalculate and apply the styles to expand or collapse the expander
-	 * according to its current state.
-	 * @param {Boolean} isSilent [false] Set to true to avoid firing the `oExpander.expand` or `oExpander.collapse` events.
+	 * Construct a custom expander. Useful to add customised expander
+	 * functionality to a component. E.g. to animate away collapsed items
+	 * rather than hide them immediately.
+	 *
+	 * @param {HTMLElement} [oExpanderElement] - The expander element in the DOM
+	 * @param {Object} [opts={}] - An options object for configuring the component, including custom class names.
 	 */
-	apply (isSilent) {
-		if (!this._isActive()) {
-			this.oExpanderElement.classList.add('o-expander--inactive');
-		} else {
-			//Remove the inactive class, this expander may be toggled.
-			this.oExpanderElement.classList.remove('o-expander--inactive');
-			// Mark collapsible items with the `o-expander__collapsible-item` class.
-			if (typeof this.options.shrinkTo === 'number') {
-				const allCountElements = this.oExpanderElement.querySelectorAll(this.options.countSelector);
-				const collapsibleCountElements = Array.from(allCountElements).splice(this.options.shrinkTo);
-				collapsibleCountElements.forEach(el => el.classList.add('o-expander__collapsible-item'));
-			}
-			// Collapse or expand.
-			if (this.isCollapsed()) {
-				this.collapse(isSilent);
-			} else {
-				this.expand(isSilent);
-			}
-		}
+	static createCustom(oExpanderElement, opts) {
+		return new ExpanderUtility(oExpanderElement, opts);
 	}
 
 	/**
-	 * Toggle the expander so expands or, if it's already expanded, collapses.
+	 * Initialise the component.
+	 * @param {(HTMLElement|String)} rootElement - The root element to initialise the component in, or a CSS selector for the root element
+	 * @param {Object} [opts={}] - An options object for configuring the component
+	 * @returns {(Expander|Array<Expander>)} - Expander instance(s)
 	 */
-	toggle () {
-		if (this.isCollapsed()) {
-			this.expand();
-		} else {
-			this.collapse();
+	static init(rootEl, opts) {
+		if (!rootEl) {
+			rootEl = document.body;
 		}
-	}
-
-	/**
-	 * Expand the expander.
-	 * @param {Boolean} isSilent [false] Set to true to avoid firing the `oExpander.expand` event.
-	 */
-	expand (isSilent) {
-		this._setExpandedState('expand', isSilent);
-	}
-
-	/**
-	 * Collapse the expander.
-	 * @param {Boolean} isSilent [false] Set to true to avoid firing the `oExpander.collapse` event.
-	 */
-	collapse (isSilent) {
-		this._setExpandedState('collapse', isSilent);
-	}
-
-	/**
-	 * Return true if the expander is currently collapse.
-	 * @returns {Boolean}
-	 */
-	isCollapsed () {
-		// If the expander has been run we store the current state.
-		if (this._currentState) {
-			return this._currentState === 'collapse';
+		if (!(rootEl instanceof HTMLElement)) {
+			rootEl = document.querySelector(rootEl);
 		}
-		// If not check for dom attributes to decide if the user intends
-		// the expander to be expanded or collapsed by default.
-		if (this.options.shrinkTo === 'hidden') {
-			return this.contentElement.getAttribute('aria-hidden') === 'true';
+		if (rootEl instanceof HTMLElement && rootEl.matches('[data-o-component=o-expander]')) {
+			return new Expander(rootEl, opts);
 		}
-		return !this.contentElement.classList.contains('o-expander--expanded');
-	}
-
-	/**
-	 * Remove the expander from the page.
-	 */
-	destroy() {
-		if (this.options.shrinkTo === 'height') {
-			document.body.removeEventListener('oViewport.orientation', () => this.apply());
-			document.body.removeEventListener('oViewport.resize', () => this.apply());
-		}
-		this.toggles.forEach(toggle => {
-			toggle.removeEventListener('click', this.toggle);
-			toggle.removeAttribute('aria-controls');
-			toggle.removeAttribute('aria-expanded');
-		});
-		this.contentElement.removeAttribute('aria-hidden');
-		this.contentElement.classList.remove('o-expander--expanded');
-		this.contentElement.classList.remove('o-expander--collapsed');
-		this.oExpanderElement.removeAttribute('data-o-expander-js');
-	}
-
-	/**
-	 * Return whether the expander has something to hide / show.
-	 * i.e. if expanding/collapsing would do anything.
-	 * @returns {Boolean}
-	 * @access private
-	 */
-	_isActive() {
-		// An expander may always toggle an expander which hides items.
-		if (this.options.shrinkTo === 'hidden') {
-			return true;
-		}
-		// An expander based on the number of items in a container may only
-		// collapse if the items length exceeds the number to shrink to. I.e.
-		// a list of 2 can't collapse to 5.
-		if (typeof this.options.shrinkTo === 'number') {
-			const expandableElements = this.oExpanderElement.querySelectorAll(this.options.countSelector);
-			if (expandableElements.length > this.options.shrinkTo) {
-				return true;
-			}
-		}
-		// If the expander is based on a height then check the content overflows
-		// the content container.
-		let overflows = false;
-		if (this.isCollapsed()) {
-			overflows = this.contentElement.clientHeight < this.contentElement.scrollHeight;
-		} else {
-			this.collapse();
-			overflows = this.contentElement.clientHeight < this.contentElement.scrollHeight;
-			this.expand();
-		}
-		return overflows;
-	}
-
-	/**
-	 * Expand or collapse the expander.
-	 * @param {Boolean} state "expand" or "collapse".
-	 * @param {Boolean} isSilent [false] Set to true to avoid firing the `oExpander.collapse` or `oExpander.expand` events.
-	 * @access private
-	 */
-	_setExpandedState(state, isSilent) {
-		// Record the current state of the expander.
-		this._currentState = state;
-		// If not hiding elements set the expanded and collapsed classes.
-		if (this.options.shrinkTo !== 'hidden') {
-			this.contentElement.classList.toggle('o-expander--expanded', state === 'expand');
-			this.contentElement.classList.remove('o-expander--collapsed', state !== 'expand');
-		}
-		// If hiding elements set `aria-hidden`.
-		if (this.options.shrinkTo === 'hidden') {
-			const value = state === 'expand' ? 'false' : 'true';
-			this.contentElement.setAttribute('aria-hidden', value);
-		}
-		// Set the toggle text and `aria-expanded` attribute.
-		if (this.options.toggleState !== 'none') {
-			this.toggles.forEach(toggle => {
-				if (this.options.toggleState !== 'aria') {
-					toggle.innerHTML = (state === 'expand' ?
-						this.options.expandedToggleText :
-						this.options.collapsedToggleText) + '<i></i>';
-				}
-				toggle.setAttribute('aria-expanded', state === 'expand' ? 'true' : 'false');
-			});
-		}
-		// Dispatch `oExpander.collapse` or `oExpander.expand` event.
-		if (!isSilent) {
-			this._dispatchEvent(state);
-		}
-	}
-
-	/**
-	 * Fire a bubbling o-expander event with the correct namespace.
-	 * @param {string} name The event name. E.g. "example" will fire an "oExpander.example" event.
-	 * @access private
-	 */
-	_dispatchEvent (name) {
-		this.oExpanderElement.dispatchEvent(new CustomEvent('oExpander.' + name, { bubbles: true }));
+		return Array.from(rootEl.querySelectorAll('[data-o-component="o-expander"]'), rootEl => new Expander(rootEl, opts));
 	}
 
 	/**
@@ -278,7 +62,7 @@ class Expander {
 	 * @returns {Object} - Data attributes as an object
 	 * @access private
 	 */
-	static _getDataAttributes (oExpanderElement) {
+	static _getDataAttributes(oExpanderElement) {
 		if (!(oExpanderElement instanceof HTMLElement)) {
 			return {};
 		}
@@ -298,25 +82,6 @@ class Expander {
 			}
 			return options;
 		}, {});
-	}
-
-	/**
-	 * Initialise the component.
-	 * @param {(HTMLElement|String)} rootElement - The root element to initialise the component in, or a CSS selector for the root element
-	 * @param {Object} [opts={}] - An options object for configuring the component
-	 * @returns {(Expander|Array<Expander>)} - Expander instance(s)
-	 */
-	static init (rootEl, opts) {
-		if (!rootEl) {
-			rootEl = document.body;
-		}
-		if (!(rootEl instanceof HTMLElement)) {
-			rootEl = document.querySelector(rootEl);
-		}
-		if (rootEl instanceof HTMLElement && rootEl.matches('[data-o-component=o-expander]')) {
-			return new Expander(rootEl, opts);
-		}
-		return Array.from(rootEl.querySelectorAll('[data-o-component="o-expander"]'), rootEl => new Expander(rootEl, opts));
 	}
 }
